InitializeKeyspace doesnt care about fuse
